update gradle build for faster deploys
When deploying to remote server, we now send over the war without any
libs (1-2MB). Then unpack the war, copy the dependency jars from the
remote machine. Then repack the war and copy to the webapps directory.

This is done to speed up the redeploy process as project dependencies
are not update often. With this method we go from a transfering a 50MB
war to a 1.5MB war