Fixed some bugs with fastqMetadata validate & upload
Added some temp code in order to add the parsed fastq filenames to the
processController.fastqMetadata object. This will be refactored when we
are able to directly pass the FastqMetadata object to FastqFileManager
instead of the JSON representation.

I think that we need to refactor the expedition table to be in the
fims-implemetation instead of currently in the Bcid table.