Fixing QueryController bugs
Fixing a bug where the "expeditions" string was changed to
"expeditionsString" in the POSTQueryResults method.

Need to send projectId along with es query endpoint.

Removing the regexp query in place of a match: {"_all": ""} query. We
updated elasticsearch tokenization & filters to be more lenient. ES
regexp queries are resource intensive and we don't want to support
those.