Remove swagger ui from war file.
We now are serving the swagger ui as static file directly from jetty.
This allows us to keep the swagger ui files directly on the server, thus
we don't need to deploy them when we deploy fims, as they rarely change.
This also allows us to use the same swagger ui across all fims
instances.

Also add the apiVersions list to the generated service.json. This will
allow us to easily switch between apiversion in the swagger docs.

The latest docs folder is called current. If we have a new api version,
we copy the apidocs from current into the previous version dir, then we
place the lastest docs into the current dir.

updated the apidocs link to point to the new swagger location.