Roll back fims-commons-version
We started to update the QueryController to use the new QueryDsl in
fims-commons v0.2.0, however we never finished and just commented out
some code. This disabled the query filters.

We are not ready to update to the new QueryDsl as we are in the process
of removing the elasticsearch dependency in commons and should wait to
update biscicol until that is finished.