Added HashMigrator
Previously hashes were not calculated correctly. keys w/o values were
being used in the calculated hash, but this is not correct and would
lead to potential duplicate records (uploading a 2nd record w/ the same
data, but an additional column on the spreadsheet w/o a value) even
though the data is the same.